feat: adds get my loan use case
