+++ src/application/modules/accounts/entities/account.ts
@@ -1,20 +1,5 @@
 import { Entity, IEntityProps } from "@/application/shared/entities/entity";
-
-export enum Roles {
-  ADMIN = 'ADMIN',
-  MANAGER = 'MANAGER',
-  BOOK_MANAGER = 'BOOK_MANAGER',
-  USER_MANAGER = 'USER_MANAGER',
-  USER = 'USER',
-}
-
-export const roleHierarchy: Record<Roles, Roles[]> = {
-  ADMIN: Object.values(Roles),
-  MANAGER: [Roles.BOOK_MANAGER, Roles.USER_MANAGER, Roles.USER],
-  USER_MANAGER: [Roles.USER],
-  BOOK_MANAGER: [],
-  USER: [],
-};
+import { Roles } from "./role";
 
 interface IAccountProps extends IEntityProps {
   name: string;
