+++ src/application/modules/accounts/entities/role.ts
@@ -1,5 +1,20 @@
 import { Entity, IEntityProps } from '@/application/shared/entities/entity';
-import { Roles } from './account';
+
+export enum Roles {
+  ADMIN = 'ADMIN',
+  MANAGER = 'MANAGER',
+  BOOK_MANAGER = 'BOOK_MANAGER',
+  USER_MANAGER = 'USER_MANAGER',
+  USER = 'USER',
+}
+
+export const roleHierarchy: Record<Roles, Roles[]> = {
+  ADMIN: Object.values(Roles),
+  MANAGER: [Roles.BOOK_MANAGER, Roles.USER_MANAGER, Roles.USER],
+  USER_MANAGER: [Roles.USER],
+  BOOK_MANAGER: [],
+  USER: [],
+};
 
 export interface RoleProps extends IEntityProps {
   code: Roles;
