+++ src/application/modules/accounts/use-cases/update-account/update-account-dto.ts
@@ -1,6 +1,6 @@
 import { generateSchema } from "@anatine/zod-openapi";
 import { z } from "zod";
-import { Roles } from "../../entities/account";
+import { Roles } from "../../entities/role";
 
 export const updateAccountSchema = z.object({
   name: z.string().min(1, 'Nome é obrigatório').optional(),
