+++ src/application/modules/loans/docs/get-my-loans-swagger.ts
@@ -1,10 +1,10 @@
 import { Tags } from "@/@types/tags";
 import { Operation } from "swagger-jsdoc";
 
-export const getLoansSwagger: Operation = {
+export const getMyLoansSwagger: Operation = {
   tags: [Tags.LOANS],
-  summary: 'Lista todos os meus empréstimos',
-  description: 'Retorna uma lista com todos os empréstimos cadastrados',
+  summary: 'Lista todos os empréstimos associados a minha conta',
+  description: 'Retorna uma lista com todos os empréstimos associados a minha conta',
   responses: {
     '200': {
       description: 'Lista de empréstimos retornada com sucesso',
