+++ src/application/modules/loans/repositories/loan-repository.ts
@@ -2,6 +2,7 @@ import { Loan } from "../entities/loan";
 
 export interface LoanRepository {
   getLoans(): Promise<Loan[]>;
+  getLoansByAccountId(accountId: string): Promise<Loan[]>;
   getLoanById(loanId: string): Promise<Loan | null>;
   createLoan(loan: Loan): Promise<void>;
   updateLoan(loan: Loan): Promise<void>;
