+++ src/application/modules/loans/use-cases/get-loans/factories/make-get-my-loans-controller.ts
@@ -1,0 +1,8 @@
+import { GetMyLoansController } from "../get-my-loans-controller";
+import { makeGetLoansUseCase } from "./make-get-loans-use-case";
+
+export function makeGetMyLoansController() {
+  const useCase = makeGetLoansUseCase();
+
+  return new GetMyLoansController(useCase);
+}
