+++ src/application/shared/http/middlewares/factories/make-authorization-middleware.ts
@@ -1,6 +1,6 @@
 
 
-import { Roles } from '@/application/modules/accounts/entities/account';
+import { Roles } from '@/application/modules/accounts/entities/role';
 import { AuthorizationMiddleware } from '../authorization-middleware';
 
 export function makeAuthorizationMiddleware(allowedRoles: Exclude<Roles, Roles.ADMIN>[]) {
