+++ src/main/express/routes/loan-router.ts
@@ -4,7 +4,8 @@ import { makeGetLoansController } from "@/application/modules/loans/use-cases/ge
 import { Router } from "express";
 import { routeAdapter } from "../adapters/route-adapter";
 
-import { Roles } from "@/application/modules/accounts/entities/account";
+import { Roles } from "@/application/modules/accounts/entities/role";
+import { makeGetMyLoansController } from "@/application/modules/loans/use-cases/get-loans/factories/make-get-MY-loans-controller";
 import { makeAuthorizationMiddleware } from "@/application/shared/http/middlewares/factories/make-authorization-middleware";
 import { middlewareAdapter } from "../adapters/middleware-adapter";
 
@@ -14,6 +15,7 @@ const authorizationMiddleware =
   middlewareAdapter(makeAuthorizationMiddleware([Roles.MANAGER, Roles.USER_MANAGER]));
 
 loanRouter.get('/', authorizationMiddleware, routeAdapter(makeGetLoansController()));
+loanRouter.get('/me', routeAdapter(makeGetMyLoansController()));
 loanRouter.post('/', authorizationMiddleware, routeAdapter(makeCreateLoanController()));
 loanRouter.patch('/:loanId', authorizationMiddleware, routeAdapter(makeCloseLoanController()));
